chapter_4/graphql_project/backend/requirements.txt
django
django-ninja
channels
channels-redis
Pillow
mysqlclient
django-cors-headers
uvicorn[standard]
strawberry-graphql-django
starlette